Refactor dashboard-week content into week-glance component
This will allow it to be re-used easily on a weekly review page.
