+++ app/templates/components/dashboard-week.hbs
@@ -1,43 +1,1 @@
-<h3>{{await title}} {{t 'general.weekAtAGlance'}}</h3>
-{{#if (is-fulfilled publishedWeekEvents)}}
-  {{#if (gt (get (await publishedWeekEvents) 'length') 0)}}
-    {{#each (await publishedWeekEvents) as |event|}}
-      <div class='event'>
-        <h4>
-          <span class='title'>{{event.name}}</span>
-          <span class='date'>
-            {{#if event.ilmSession}}
-              <span class='ilm-due'>{{t 'general.dueBy'}}</span>
-            {{/if}}
-            {{moment-format event.startDate 'dddd, MMMM Do, YYYY h:mma'}}
-          </span>
-        </h4>
-        <p class ='sessiontype'>
-          {{event.sessionTypeTitle}}
-          {{#if event.courseExternalId}}
-            - {{event.courseExternalId}}
-          {{/if}}
-        </p>
-        <p class='description'>{{event.sessionDescription}}</p>
-        <ul class='learning-materials'>
-          {{#each event.learningMaterials as |lm|}}
-            <li class='learning-material'>
-              {{#if lm.absoluteFileUri}}
-                <a href={{lm.absoluteFileUri}} target='_blank' rel='noopener'> {{fa-icon 'external-link-square'}} {{lm.title}}</a>
-              {{else if lm.link}}
-                <a href={{lm.link}} target='_blank' rel='noopener'>{{fa-icon 'external-link-square'}} {{lm.title}}</a>
-              {{else}}
-                {{lm.title}}
-                <ul><li><small>{{lm.citation}}</small></li></ul>
-              {{/if}}
-            </li>
-          {{/each}}
-        </ul>
-      </div>
-    {{/each}}
-  {{else}}
-    <p>{{t 'general.none'}}</p>
-  {{/if}}
-{{else}}
-  <p>{{fa-icon 'spinner' spin=true}} {{t 'general.loadingEvents'}}</p>
-{{/if}}
+{{week-glance year=(moment-format today 'YYYY') week=(moment-format today 'W')}}
